requirements.txt
requests
qrcode
zbarlight
Pillow
